receive/requirements.txt
pyzmq==19.0.1
pycrypto==2.6.1